Speed up actions by using a single linear layer for categorical and one for continuous. Then using batches in Categorical and Normal sampling.

+++ src/main/java/com/tenbitmelon/machinelearningplayer/environment/MinecraftEnvironment.java
@@ -422,7 +422,7 @@ public class MinecraftEnvironment {
         float damageDealt = lastKnownTargetHealth - targetHealth;
         lastKnownTargetHealth = targetHealth;
 
-        double distanceTo = agent.position().distanceTo(targetEntity.position());
+        float distanceTo = (float) agent.position().distanceTo(targetEntity.position());
         if (lastKnownDistanceToTarget == 0.0f) {
             lastKnownDistanceToTarget = (float) distanceTo;
         }
